+++ MyCad/Source/Engine/VertexBuffer.h
@@ -7,7 +7,7 @@
 class VertexBuffer
 {
 private:
-    int count;
+    unsigned int count;
     unsigned int id;
 
 public:
@@ -18,7 +18,7 @@ public:
     void Bind() const;
     void Unbind() const;
 
-    int GetCount() const;
+    unsigned int GetCount() const;
     template <typename T> void SetVertices(const std::vector<T>& vertices);
 };
 
@@ -32,7 +32,7 @@ VertexBuffer::VertexBuffer(const std::vector<T>& vertices)
 template <typename T>
 void VertexBuffer::SetVertices(const std::vector<T>& vertices)
 {
-    count = vertices.size();
+    count = static_cast<unsigned int>(vertices.size());
     Bind();
     GLCall(glBufferData(GL_ARRAY_BUFFER, vertices.size() * sizeof(T), vertices.data(), GL_DYNAMIC_DRAW));
     Unbind();
